Finished MVC Project (frist attempt)
Created a simple example of the MVC coding pattern that outputs a vector to which the user can move an "X" to the left or right. The console will be update via the view after each command.

Note: My intention was to do this in a GUI; however, I couldn't not get it to work before realizing I was spending too much time on the project. Thus I switched over to a console version.