Finished Servlets Example
I'm using the CSS file from a personal family server I am working on; however, it is built using PHP, so this was a good chance to try and translate everything over. Granted, the servlet is simple, but it none the less calcualtes wither the answer to the riddel contained the correct work.

For credit, the riddle is from The Hobiit.