Completed JSON example
This one wasn't too bad - though I admit that I perhaps made it too complex. I used the older JSON libraries instead of going with Jackson and I believe that may have made it harder to both code and find help online. Also, I tried encoding info that only the program would recognize and decode in the case of "Character". It is always only a two letter string in the JSON because each letter represents a race and class. Not needed, and certainly meant I had to code more to interpret it, but it was kind of fun to try out.