Completed System Test & updated Use Case
The System test document is created based heavily on the book's template; however, I decided to use actual screenshots within the result coulumn to ensure accuracy.

The change in the Test Case document was switching "click" to "type" for user input, as "click" is not how the use inputs thier choices.