Finished ACP and HttpURLConnections
Finished both Application Controller Pattern and HttpURLConnection example programs. The ACP replicates my Java Collections program, so the only actually "new" code was pasting methods into thier own handler. The ApplicationController class and Handler interface are heavily taken from the books example. It seem appropriate given that thier functions are so simple. All that was changed was the parameters being shorted since not data needed to be pass through, only the command.

The HttpUrlConnection example doesn't not use TomCat simply because I feared the continuel crashing I had experienced coding the servlet. Therefore, I found a website that responds to HttpURL requests and  sends a JSON or Text data and used that as the server side function.