System Level Testing and Web App .war file added
Was able to complete the System Level Testing inclusion with some extra time.

Also including the web apps .war file that I was attempting to use with Tomcat on my raspberry pi. Unfortunately I have not gotten public access to work with the purchased domain "capturescabs.com"; however, I will contintue to get it to work for my own knowledge.